error handling

+++ src/main/kotlin/data/repository/QuizQuestionRepositoryImpl.kt
@@ -2,6 +2,7 @@ package org.aprikot.data.repository
 
 import com.mongodb.client.model.Filters
 import com.mongodb.client.model.Updates
+import com.mongodb.client.result.InsertManyResult
 import com.mongodb.kotlin.client.coroutine.MongoDatabase
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
@@ -12,6 +13,8 @@ import org.aprikot.data.mapper.toQuizQuestionEntity
 import org.aprikot.data.utils.Constant.QUESTIONS_COLLECTION_NAME
 import org.aprikot.domain.model.QuizQuestion
 import org.aprikot.domain.repository.QuizQuestionRepository
+import org.aprikot.domain.util.DataError
+import org.aprikot.domain.util.Result
 import java.util.Date
 
 class QuizQuestionRepositoryImpl(
@@ -21,8 +24,8 @@ class QuizQuestionRepositoryImpl(
     val questionCollection = mongoDatabase
         .getCollection<QuizQuestionEntity>(QUESTIONS_COLLECTION_NAME)
 
-    override suspend fun upsertQuestion(question: QuizQuestion) {
-        try{
+    override suspend fun upsertQuestion(question: QuizQuestion): Result<Unit, DataError> {
+        return try{
             if (question.id == null){
                 questionCollection.insertOne(question.toQuizQuestionEntity())
             } else {
@@ -38,50 +41,76 @@ class QuizQuestionRepositoryImpl(
                     Updates.set(QuizQuestionEntity::topicCode.name, question.topicCode),
                     Updates.set(QuizQuestionEntity::updatedAt.name, Date()),
                 )
-                questionCollection.updateOne(filterQuery, updateQuery)
+
+                val updateResult = questionCollection.updateOne(filterQuery, updateQuery)
+
+                if (updateResult.modifiedCount == 0L) {
+                    return Result.Failure(DataError.NotFound)
+                }
+
             }
-        } catch (e: Exception){
+
+            Result.Success(Unit)
+
+        } catch (e: Exception) {
             e.printStackTrace()
+            Result.Failure(DataError.Database)
         }
     }
 
-    override suspend fun upsertMultipleQuestions(listOfQuestions: List<QuizQuestion>){
-        try{
+    override suspend fun upsertMultipleQuestions(listOfQuestions: List<QuizQuestion>): Result<Unit, DataError>{
+        return try{
             val listQuestionsEntity: List<QuizQuestionEntity> = listOfQuestions
                 .map { it.toQuizQuestionEntity() }
                 .toList()
-            questionCollection.insertMany(
+            val updateResult = questionCollection.insertMany(
                 listQuestionsEntity
             )
 
+            if (!updateResult.wasAcknowledged()) {
+                return Result.Failure(DataError.NotFound)
+            }
+
+            Result.Success(Unit)
+
         } catch (e: Exception){
             e.printStackTrace()
+            Result.Failure(DataError.Database)
         }
     }
 
     override suspend fun getAllQuestions(
         topicCode: Int?,
         limit: Int?
-    ): List<QuizQuestion> {
+    ): Result<List<QuizQuestion>, DataError> {
        return try {
            val questionLimit = limit?.takeIf { it > 0 } ?: 20
            val filterQuery = topicCode?.let {
                Filters.eq(QuizQuestionEntity::topicCode.name, it)
            } ?: Filters.empty()
 
-           questionCollection
+           val questions = questionCollection
                .find(filter = filterQuery)
                .limit(questionLimit)
                .map { it.toQuizQuestion() }
                .toList()
 
+           if (questions.isNotEmpty()){
+               Result.Success(questions)
+           } else {
+               Result.Failure(DataError.NotFound)
+           }
        } catch (e: Exception){
            e.printStackTrace()
-           emptyList()
+           Result.Failure(DataError.Database)
        }
     }
 
-    override suspend fun getQuestionById(id: String): QuizQuestion? {
+    override suspend fun getQuestionById(id: String?): Result<QuizQuestion, DataError> {
+        if (id.isNullOrBlank()){
+            return Result.Failure(DataError.Validation)
+        }
+
         return try{
             val filterQuery = Filters.eq(
                 QuizQuestionEntity::_id.name, id
@@ -89,24 +118,40 @@ class QuizQuestionRepositoryImpl(
             val questionEntity = questionCollection
                 .find(filterQuery)
                 .firstOrNull()
-            questionEntity?.toQuizQuestion()
+            if (questionEntity != null){
+                val question = questionEntity.toQuizQuestion()
+                Result.Success(question)
+            } else {
+                Result.Failure(DataError.NotFound)
+            }
+
         } catch (e: Exception){
             e.printStackTrace()
-            null
+            Result.Failure(DataError.Database)
         }
     }
 
-    override suspend fun deleteQuestionById(id: String): Boolean {
+    override suspend fun deleteQuestionById(id: String?): Result<Unit, DataError> {
+        if (id.isNullOrBlank()){
+            return Result.Failure(DataError.Validation)
+        }
+
         return try {
             val filterQuery = Filters.eq(
                 QuizQuestionEntity::_id.name, id
             )
+
             val deleteResult = questionCollection
                 .deleteOne(filter = filterQuery)
-            deleteResult.deletedCount > 0
+
+            if (deleteResult.deletedCount > 0){
+                Result.Success(Unit)
+            } else {
+                Result.Failure(DataError.NotFound)
+            }
         } catch (e: Exception){
             e.printStackTrace()
-            false
+            Result.Failure(DataError.Database)
         }
     }
